requirements-client.txt
google-cloud-storage
python-dotenv
requests
